+++ types/tests/Mock.ts
@@ -1,0 +1,5 @@
+type Mock<I, O> = {
+	input: I;
+	output: O;
+	description?: string;
+};