+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # vercel-ci-cd-with-changesets
 
+## 0.0.3
+
+### Patch Changes
+
+- b2cd56d: Added a Counter component with a decrement and an increment button, along with an initialCount prop and a reset button
+  Added two Counters to the homepage
+
 ## 0.0.2
 
 ### Patch Changes